A0 writeup truly finished

+++ tests/a0/test.a0.task2.problems.cpp
@@ -41,7 +41,7 @@ Test test_a0_task2_problems_numerical("a0.task2.problems.numerical", []() {
     // Ex) let x = 1, y = 4. y / 3 = 4 / 3 = 1.333 > 1, so x < y / 3. Return true.
 
     int factor = 3;
-    auto helper = [&](int x, int y) { return static_cast<float>(x) < (static_cast<float>(y) / static_cast<float>(factor)); }; // integer divion issue
+    auto helper = [&](int x, int y) { return x < (static_cast<float>(y) / factor); }; // integer divion issue
 
     int j = 0;
     for (auto& v : target) {
